refactor(html): extract SOBR, repo, and job tables from god class
Extract 7 table classes from CHtmlTables.cs (3793 -> 2144 lines, 43% reduction):
- CSobrTable, CSobrExtentTable: SOBR and extent tables
- CCapacityTierExtTable, CArchiveTierExtTable: tier extent tables
- CRepoTable: standalone repository table
- CJobSessionSummaryTable: job session summary (flat + by-job views)
- CJobInfoTable: job info with tape/entra sub-tables
