+++ vHC/HC_Reporting/Functions/Reporting/Html/VBR/VbrTables/CHtmlTables.cs
@@ -24,6 +24,7 @@ using VeeamHealthCheck.Functions.Reporting.Html.VBR.VbrTables.TapeInfra;
 using VeeamHealthCheck.Functions.Reporting.Html.VBR.VbrTables.CloudConnect;
 using VeeamHealthCheck.Functions.Reporting.Html.VBR.VbrTables.SureBackup;
 using VeeamHealthCheck.Functions.Reporting.Html.VBR.VbrTables.GeneralSettings;
+using VeeamHealthCheck.Functions.Reporting.Html.VBR.VbrTables.SOBR;
 using VeeamHealthCheck.Reporting.Html.VBR;
 using VeeamHealthCheck.Resources.Localization;
 using VeeamHealthCheck.Scrubber;
@@ -348,6 +349,14 @@ namespace VeeamHealthCheck.Html.VBR
 
         // helper to set a generic section in JSON aggregation
         private static void SetSection(string key, List<string> headers, List<List<string>> rows, string summary)
+        {
+            SetSectionPublic(key, headers, rows, summary);
+        }
+
+        /// <summary>
+        /// Public accessor for SetSection, used by extracted table classes.
+        /// </summary>
+        internal static void SetSectionPublic(string key, List<string> headers, List<List<string>> rows, string summary)
         {
             if (CGlobals.FullReportJson == null)
             {
@@ -1639,666 +1648,32 @@ namespace VeeamHealthCheck.Html.VBR
 
         public string AddSobrTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("sobr", VbrLocalizationHelper.SbrTitle, VbrLocalizationHelper.SbrBtn);
-            string summary = this.sum.Sobr();
-            s += "<tr>" +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr0, VbrLocalizationHelper.Sbr0TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr1, VbrLocalizationHelper.Sbr1TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Repo0, VbrLocalizationHelper.Repo1TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr2, VbrLocalizationHelper.Sbr2TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr3, VbrLocalizationHelper.Sbr3TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr4, VbrLocalizationHelper.Sbr4TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr5, VbrLocalizationHelper.Sbr5TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr6, VbrLocalizationHelper.Sbr6TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr7, VbrLocalizationHelper.Sbr7TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr8, VbrLocalizationHelper.Sbr8TT) +
-           this.form.TableHeader("CapTier Immutable", VbrLocalizationHelper.Sbr9TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr10, VbrLocalizationHelper.Sbr10TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr11, VbrLocalizationHelper.Sbr11TT) +
-           this.form.TableHeader(VbrLocalizationHelper.Sbr12, VbrLocalizationHelper.Sbr12TT) +
-           "</tr>";
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-
-            // CDataFormer cd = new(true);
-            try
-            {
-                this.log.Info("Attempting to load SOBR data...");
-                List<CSobrTypeInfos> list = this.df.SobrInfoToXml(scrub);
-
-                if (list == null || list.Count == 0)
-                {
-                    this.log.Warning("No SOBR data found. The SOBRs CSV file may be missing or empty.");
-                    this.log.Info("This could indicate: 1) No SOBRs configured, 2) Collection script failed, or 3) CSV file not found");
-                }
-
-                foreach (var d in list)
-                {
-                    s += "<tr>";
-                    s += this.form.TableData(d.Name, string.Empty);
-                    s += this.form.TableData(d.ExtentCount.ToString(), string.Empty);
-                    s += this.form.TableData(d.JobCount.ToString(), string.Empty);
-                    s += this.form.TableData(d.PolicyType, string.Empty);
-                    if (d.EnableCapacityTier)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.CapacityTierCopyPolicyEnabled)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.CapacityTierMovePolicyEnabled)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.ArchiveTierEnabled)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.UsePerVMBackupFiles)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.CapTierType, string.Empty);
-                    if (d.ImmuteEnabled)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.ImmutePeriod, string.Empty);
-                    if (d.SizeLimitEnabled)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.SizeLimit, string.Empty);
-                    s += "</tr>";
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("SOBR Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON SOBR
-            try
-            {
-                var list = this.df.SobrInfoToXml(scrub) ?? new List<CSobrTypeInfos>();
-                List<string> headers = new() { "Name", "ExtentCount", "JobCount", "PolicyType", "EnableCapacityTier", "CapacityTierCopy", "CapacityTierMove", "ArchiveTierEnabled", "UsePerVMFiles", "CapTierType", "ImmutableEnabled", "ImmutablePeriod", "SizeLimitEnabled", "SizeLimit" };
-                List<List<string>> rows = list.Select(d => new List<string>
-                {
-                    d.Name,
-                    d.ExtentCount.ToString(),
-                    d.JobCount.ToString(),
-                    d.PolicyType,
-                    d.EnableCapacityTier ? "True" : "False",
-                    d.CapacityTierCopyPolicyEnabled ? "True" : "False",
-                    d.CapacityTierMovePolicyEnabled ? "True" : "False",
-                    d.ArchiveTierEnabled ? "True" : "False",
-                    d.UsePerVMBackupFiles ? "True" : "False",
-                    d.CapTierType,
-                    d.ImmuteEnabled ? "True" : "False",
-                    d.ImmutePeriod,
-                    d.SizeLimitEnabled ? "True" : "False",
-                    d.SizeLimit,
-                }).ToList();
-                SetSection("sobr", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture sobr JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CSobrTable();
+            return table.Render(scrub);
         }
 
         public string AddSobrExtTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("extents", VbrLocalizationHelper.SbrExtTitle, VbrLocalizationHelper.SbrExtBtn);
-            string summary = this.sum.Extents();
-            s += "<tr>" +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt0, VbrLocalizationHelper.SbrExt0TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt1, VbrLocalizationHelper.SbrExt1TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt2, VbrLocalizationHelper.SbrExt2TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt3, VbrLocalizationHelper.SbrExt3TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt4, VbrLocalizationHelper.SbrExt4TT) +
-this.form.TableHeader("Auto Gateway / Direct Connection", VbrLocalizationHelper.SbrExt5TT) +
-this.form.TableHeader("Specified Gateway(s)", VbrLocalizationHelper.SbrExt6TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt7, VbrLocalizationHelper.SbrExt7TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt8, VbrLocalizationHelper.SbrExt8TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt9, VbrLocalizationHelper.SbrExt9TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt10, VbrLocalizationHelper.SbrExt10TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt11, VbrLocalizationHelper.SbrExt11TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt12, VbrLocalizationHelper.SbrExt12TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt13, VbrLocalizationHelper.SbrExt13TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt14, VbrLocalizationHelper.SbrExt14TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrExt15TT) +
-"</tr>";
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-            try
-            {
-                this.log.Info("Attempting to load SOBR Extent data...");
-                List<CRepository> list = this.df.ExtentXmlFromCsv(scrub);
-
-                if (list == null || list.Count == 0)
-                {
-                    this.log.Warning("No SOBR Extent data found. The SOBRExtents CSV file may be missing or empty.");
-                    this.log.Info("This could indicate: 1) No SOBRs configured, 2) Collection script failed, or 3) CSV file not found");
-                }
-
-                foreach (var d in list)
-                {
-                    var prov = d.Provisioning;
-                    int shade = 0;
-                    if (prov == "under")
-                    {
-                        shade = 2;
-                    }
-
-                    if (prov == "over")
-                    {
-                        shade = 1;
-                    }
-
-                    int freeSpaceShade = 0;
-
-                    // decimal.TryParse(d.FreeSpacePercent, out decimal i);
-                    if (d.FreeSpacePercent < 20) { freeSpaceShade = 1; }
-
-                    s += "<tr>";
-                    s += this.form.TableData(d.Name, string.Empty);
-                    s += this.form.TableData(d.SobrName, string.Empty);
-                    s += this.form.TableData(d.MaxTasks.ToString(), string.Empty, shade);
-                    s += this.form.TableData(d.Cores.ToString(), string.Empty);
-                    s += this.form.TableData(d.Ram.ToString(), string.Empty);
-
-                    if (d.IsAutoGate)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.Host, string.Empty);
-                    s += this.form.TableData(d.Path, string.Empty);
-                    s += this.form.TableData(d.FreeSpace.ToString(), string.Empty);
-                    s += this.form.TableData(d.TotalSpace.ToString(), string.Empty);
-                    s += this.form.TableData(d.FreeSpacePercent.ToString(), string.Empty, freeSpaceShade);
-
-                    if (d.IsDecompress)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.AlignBlocks)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.IsRotatedDrives)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.IsImmutabilitySupported)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.Type, string.Empty);
-
-                    // s += _form.TableData(d[16], "");
-                    s += "</tr>";
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Extents Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON extents
-            try
-            {
-                var list = this.df.ExtentXmlFromCsv(scrub) ?? new List<CRepository>();
-                List<string> headers = new() { "Name", "SobrName", "MaxTasks", "Cores", "Ram", "IsAutoGate", "Host", "Path", "FreeSpace", "TotalSpace", "FreeSpacePercent", "IsDecompress", "AlignBlocks", "IsRotatedDrives", "IsImmutabilitySupported", "Type" };
-                List<List<string>> rows = list.Select(d => new List<string>
-                {
-                    d.Name,
-                    d.SobrName,
-                    d.MaxTasks.ToString(),
-                    d.Cores.ToString(),
-                    d.Ram.ToString(),
-                    d.IsAutoGate ? "True" : "False",
-                    d.Host,
-                    d.Path,
-                    d.FreeSpace.ToString(),
-                    d.TotalSpace.ToString(),
-                    d.FreeSpacePercent.ToString(),
-                    d.IsDecompress ? "True" : "False",
-                    d.AlignBlocks ? "True" : "False",
-                    d.IsRotatedDrives ? "True" : "False",
-                    d.IsImmutabilitySupported ? "True" : "False",
-                    d.Type,
-                }).ToList();
-                SetSection("extents", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture extents JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CSobrExtentTable();
+            return table.Render(scrub);
         }
 
         public string AddRepoTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("repos", VbrLocalizationHelper.RepoTitle, VbrLocalizationHelper.RepoBtn);
-            string summary = this.sum.Repos();
-            s +=
-this.form.TableHeader(VbrLocalizationHelper.SbrExt0, VbrLocalizationHelper.SbrExt0TT, 0) +
-this.form.TableHeader(VbrLocalizationHelper.Repo0, VbrLocalizationHelper.Repo0TT, 1) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt2, VbrLocalizationHelper.SbrExt2TT, 2) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt3, VbrLocalizationHelper.SbrExt3TT, 3) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt4, VbrLocalizationHelper.SbrExt4TT, 4) +
-this.form.TableHeader("Auto Gateway / Direct Connection", VbrLocalizationHelper.SbrExt5TT) +
-this.form.TableHeader("Specified Gateway(s)", VbrLocalizationHelper.SbrExt6TT) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt7, VbrLocalizationHelper.SbrExt7TT, 7) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt8, VbrLocalizationHelper.SbrExt8TT, 8) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt9, VbrLocalizationHelper.SbrExt9TT, 9) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt10, VbrLocalizationHelper.SbrExt10TT, 10) +
-this.form.TableHeader(VbrLocalizationHelper.Repo1, VbrLocalizationHelper.Repo1TT, 11) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt11, VbrLocalizationHelper.SbrExt11TT, 12) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt12, VbrLocalizationHelper.SbrExt12TT, 13) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt13, VbrLocalizationHelper.SbrExt13TT, 14) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt14, VbrLocalizationHelper.SbrExt14TT, 15) +
-this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrExt15TT, 16);
-
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-            try
-            {
-                List<CRepository> list = this.df.RepoInfoToXml(scrub);
-
-                foreach (var d in list)
-                {
-                    var prov = d.Provisioning;
-                    int shade = 0;
-                    if (prov == "under")
-                    {
-                        shade = 2;
-                    }
-
-                    if (prov == "over")
-                    {
-                        shade = 1;
-                    }
-
-                    int freeSpaceShade = 0;
-
-                    // decimal.TryParse(d.FreeSpacePercent, out decimal i);
-
-                    // int perVmShade = 0;
-                    // if (d[11] == "False")
-                    //    perVmShade = 3;
-                    if (d.FreeSpacePercent < 20) { freeSpaceShade = 1; }
-                    s += "<tr>";
-                    s += this.form.TableData(d.Name, string.Empty);
-                    s += this.form.TableData(d.JobCount.ToString(), string.Empty);
-                    s += this.form.TableData(d.MaxTasks.ToString(), string.Empty, shade);
-                    s += this.form.TableData(d.Cores.ToString(), string.Empty);
-                    s += this.form.TableData(d.Ram.ToString(), string.Empty);
-                    if (d.IsAutoGate)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.Host, string.Empty);
-                    s += this.form.TableData(d.Path, string.Empty);
-                    s += this.form.TableData(d.FreeSpace.ToString(), string.Empty);
-                    s += this.form.TableData(d.TotalSpace.ToString(), string.Empty);
-                    s += this.form.TableData(d.FreeSpacePercent.ToString(), string.Empty, freeSpaceShade);
-                    if (d.IsPerVmBackupFiles)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.Warn, string.Empty);
-                    }
-
-                    if (d.IsDecompress)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.AlignBlocks)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.IsRotatedDrives)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    if (d.IsImmutabilitySupported)
-                    {
-                        s += this.form.TableData(this.form.True, string.Empty);
-                    }
-                    else
-                    {
-                        s += this.form.TableData(this.form.False, string.Empty);
-                    }
-
-                    s += this.form.TableData(d.Type, string.Empty);
-
-                    s += "</tr>";
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("REPO Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON repos
-            try
-            {
-                var list = this.df.RepoInfoToXml(scrub) ?? new List<CRepository>();
-                List<string> headers = new() { "Name", "JobCount", "MaxTasks", "Cores", "Ram", "IsAutoGate", "Host", "Path", "FreeSpace", "TotalSpace", "FreeSpacePercent", "IsPerVmBackupFiles", "IsDecompress", "AlignBlocks", "IsRotatedDrives", "IsImmutabilitySupported", "Type" };
-                List<List<string>> rows = list.Select(d => new List<string>
-                {
-                    d.Name,
-                    d.JobCount.ToString(),
-                    d.MaxTasks.ToString(),
-                    d.Cores.ToString(),
-                    d.Ram.ToString(),
-                    d.IsAutoGate ? "True" : "False",
-                    d.Host,
-                    d.Path,
-                    d.FreeSpace.ToString(),
-                    d.TotalSpace.ToString(),
-                    d.FreeSpacePercent.ToString(),
-                    d.IsPerVmBackupFiles ? "True" : "False",
-                    d.IsDecompress ? "True" : "False",
-                    d.AlignBlocks ? "True" : "False",
-                    d.IsRotatedDrives ? "True" : "False",
-                    d.IsImmutabilitySupported ? "True" : "False",
-                    d.Type,
-                }).ToList();
-                SetSection("repos", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture repos JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CRepoTable();
+            return table.Render(scrub);
         }
 
-        /// <summary>
-        /// Generates HTML table displaying Capacity Tier extents for all SOBRs
-        /// </summary>
         public string AddCapacityTierExtTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("capextents", "Capacity Tier Configuration", "Capacity");
-            string summary = "Capacity tier extent configuration and retention policies";
-            s += "<tr>" +
-                this.form.TableHeader("Name", "Name of the capacity tier extent") +
-                this.form.TableHeader("SOBR Name", "Parent Scale-Out Backup Repository") +
-                this.form.TableHeader("Status", "Capacity tier status") +
-                this.form.TableHeader("Auto Gateway / Direct Connection", "Gateway connection type (e.g., Direct, Gateway)") +
-                this.form.TableHeader("Specified Gateway(s)", "Gateway servers assigned to this capacity extent") +
-                this.form.TableHeader("Copy Mode Enabled", "Whether copy mode is enabled for capacity tier") +
-                this.form.TableHeader("Move Mode Enabled", "Whether move mode is enabled for capacity tier") +
-                this.form.TableHeader("Move Period (Days)", "Age threshold before data is moved to capacity tier") +
-                this.form.TableHeader("Encryption", "Whether encryption is enabled for capacity tier") +
-                this.form.TableHeader("Use Immutability", "Whether immutability is enforced") +
-                this.form.TableHeader("Immutable Period", "How long data is locked in capacity tier") +
-                this.form.TableHeader("Immutability Mode", "Immutability mode (v13+, e.g. RepositoryRetention)") +
-                this.form.TableHeader("Size Limit Enabled", "Whether size limiting is enforced") +
-                this.form.TableHeader("Size Limit", "Maximum capacity tier size") +
-                this.form.TableHeader("Type", "Repository type (e.g., AWS S3, Azure Blob)") +
-                "</tr>";
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-
-            List<CCapacityTierExtent> list = new();
-            try
-            {
-                this.log.Info("Attempting to load Capacity Tier Extent data...");
-                list = this.df.CapacityTierXmlFromCsv(scrub) ?? list;
-
-                if (list.Count == 0)
-                {
-                    this.log.Warning("No Capacity Tier Extent data found. No SOBRs with capacity tier may be configured.");
-                }
-
-                foreach (var d in list)
-                {
-                    s += "<tr>";
-                    s += this.form.TableData(d.Name, string.Empty);
-                    s += this.form.TableData(d.SobrName, string.Empty);
-                    s += this.form.TableData(Badge(d.Status), string.Empty);
-                    s += this.form.TableData(d.ConnectionType ?? string.Empty, string.Empty);
-                    s += this.form.TableData(d.GatewayServer ?? string.Empty, string.Empty);
-                    s += this.BoolCell(d.CopyModeEnabled);
-                    s += this.BoolCell(d.MoveModeEnabled);
-                    s += this.form.TableData(d.MovePeriodDays.ToString(), string.Empty);
-                    s += this.BoolCell(d.EncryptionEnabled);
-                    s += this.BoolCell(d.ImmutableEnabled);
-                    s += this.form.TableData(d.ImmutablePeriod, string.Empty);
-                    s += this.form.TableData(d.ImmutabilityMode ?? string.Empty, string.Empty);
-                    s += this.BoolCell(d.SizeLimitEnabled);
-                    s += this.form.TableData(d.SizeLimit ?? string.Empty, string.Empty);
-                    s += this.form.TableData(d.Type, string.Empty);
-                    s += "</tr>";
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Capacity Tier Extent Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON Capacity Tier Extents
-            try
-            {
-                List<string> headers = new() { "Name", "SobrName", "Status", "ConnectionType", "GatewayServer", "CopyModeEnabled", "MoveModeEnabled", "MovePeriodDays", "EncryptionEnabled", "ImmutableEnabled", "ImmutablePeriod", "ImmutabilityMode", "SizeLimitEnabled", "SizeLimit", "Type" };
-                List<List<string>> rows = list.Select(d => new List<string>
-                {
-                    d.Name,
-                    d.SobrName,
-                    d.Status,
-                    d.ConnectionType ?? string.Empty,
-                    d.GatewayServer ?? string.Empty,
-                    d.CopyModeEnabled ? "True" : "False",
-                    d.MoveModeEnabled ? "True" : "False",
-                    d.MovePeriodDays.ToString(),
-                    d.EncryptionEnabled ? "True" : "False",
-                    d.ImmutableEnabled ? "True" : "False",
-                    d.ImmutablePeriod,
-                    d.ImmutabilityMode ?? string.Empty,
-                    d.SizeLimitEnabled ? "True" : "False",
-                    d.SizeLimit ?? string.Empty,
-                    d.Type,
-                }).ToList();
-                SetSection("capextents", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture capacity tier extents JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CCapacityTierExtTable();
+            return table.Render(scrub);
         }
 
-        /// <summary>
-        /// Generates HTML table displaying Archive Tier configuration for all SOBRs
-        /// </summary>
         public string AddArchiveTierExtTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("archextents", "Archive Tier Configuration", "Archive");
-            string summary = "Archive tier extent retention and immutability policies";
-            s += "<tr>" +
-                this.form.TableHeader("Name", "Name of archive tier repository/bucket") +
-                this.form.TableHeader("SOBR Name", "Scale-Out Backup Repository name") +
-                this.form.TableHeader("Status", "Archive extent status") +
-                this.form.TableHeader("Auto Gateway / Direct Connection", "Gateway connection mode (e.g., Direct, Gateway)") +
-                this.form.TableHeader("Specified Gateway(s)", "Gateway servers assigned to this archive extent") +
-                this.form.TableHeader("Offload Period", "How long before data is offloaded to archive") +
-                this.form.TableHeader("Archive Tier Enabled", "Whether archive tier is active") +
-                this.form.TableHeader("Cost Optimized", "Cost optimization policy enabled") +
-                this.form.TableHeader("Full Backup Mode", "Full backup mode for archive tier") +
-                this.form.TableHeader("Encryption", "Whether encryption is enabled for archive tier") +
-                this.form.TableHeader("Use Immutability", "Whether immutability is enforced on archive") +
-                this.form.TableHeader("Type", "Archive repository type (e.g., AmazonS3, AzureBlob)") +
-                "</tr>";
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-
-            List<CArchiveTierExtent> list = new();
-            try
-            {
-                this.log.Info("Attempting to load Archive Tier Extent data...");
-                list = this.df.ArchiveTierXmlFromCsv(scrub) ?? list;
-
-                if (list.Count == 0)
-                {
-                    this.log.Warning("No Archive Tier Extent data found. No SOBRs with archive tier may be configured.");
-                }
-
-                foreach (var d in list)
-                {
-                    s += "<tr>";
-                    s += this.form.TableData(d.Name, string.Empty);
-                    s += this.form.TableData(d.SobrName, string.Empty);
-                    s += this.form.TableData(Badge(d.Status), string.Empty);
-                    s += this.form.TableData(d.GatewayMode ?? string.Empty, string.Empty);
-                    s += this.form.TableData(d.GatewayServer ?? string.Empty, string.Empty);
-                    s += this.form.TableData(d.OffloadPeriod ?? string.Empty, string.Empty);
-                    s += this.BoolCell(d.ArchiveTierEnabled);
-                    s += this.BoolCell(d.CostOptimizedEnabled);
-                    s += this.BoolCell(d.FullBackupModeEnabled);
-                    s += this.BoolCell(d.EncryptionEnabled);
-                    s += this.BoolCell(d.ImmutableEnabled);
-                    s += this.form.TableData(d.Type ?? string.Empty, string.Empty);
-                    s += "</tr>";
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Archive Tier Extent Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON Archive Tier Extents
-            try
-            {
-                List<string> headers = new() { "Name", "SobrName", "Status", "GatewayMode", "GatewayServer", "OffloadPeriod", "ArchiveTierEnabled", "CostOptimizedEnabled", "FullBackupModeEnabled", "EncryptionEnabled", "ImmutableEnabled", "Type" };
-                List<List<string>> rows = list.Select(d => new List<string>
-                {
-                    d.Name,
-                    d.SobrName,
-                    d.Status,
-                    d.GatewayMode ?? string.Empty,
-                    d.GatewayServer ?? string.Empty,
-                    d.OffloadPeriod ?? string.Empty,
-                    d.ArchiveTierEnabled ? "True" : "False",
-                    d.CostOptimizedEnabled ? "True" : "False",
-                    d.FullBackupModeEnabled ? "True" : "False",
-                    d.EncryptionEnabled ? "True" : "False",
-                    d.ImmutableEnabled ? "True" : "False",
-                    d.Type ?? string.Empty,
-                }).ToList();
-                SetSection("archextents", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture archive tier extents JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CArchiveTierExtTable();
+            return table.Render(scrub);
         }
 
         public string AddJobConTable(bool scrub)
@@ -2416,121 +1791,8 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
 
         public string AddJobSessSummTable(bool scrub)
         {
-            this.log.Info("Adding Job Session Summary Table");
-            string s = this.form.SectionStartWithButton("jobsesssum", VbrLocalizationHelper.JssTitle, VbrLocalizationHelper.JssBtn, CGlobals.ReportDays);
-            string summary = this.sum.JobSessSummary();
-
-            s += this.form.TableHeaderLeftAligned(VbrLocalizationHelper.Jss0, VbrLocalizationHelper.Jss0TT); // job name
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss1, VbrLocalizationHelper.Jss1TT);// items
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss2, VbrLocalizationHelper.Jss2TT); // min time
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss3, VbrLocalizationHelper.Jss3TT);// max time
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss4, VbrLocalizationHelper.Jss4TT);// avg time
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss5, VbrLocalizationHelper.Jss5TT); // total sessions
-            s += this.form.TableHeader("Fails", "Total times job failed"); // fails
-            s += this.form.TableHeader("Retries", "Total times job retried");// retries
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss6, VbrLocalizationHelper.Jss6TT);// success rate
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss7, VbrLocalizationHelper.Jss7TT); // avg backup size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss8, VbrLocalizationHelper.Jss8TT);// max backup size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss9, VbrLocalizationHelper.Jss9TT); // avg data size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss10, "Provisioned size of all objects in job."); // max data size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss11, "Avg Data Size divided by Max Data Size (average processed data divided by total consumed size of all VMs in the job)"); // avg change rate
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss12, VbrLocalizationHelper.Jss12TT); // wait for res count
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss13, VbrLocalizationHelper.Jss13TT); // max wait
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss14, VbrLocalizationHelper.Jss14TT);// avg wait
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss15, VbrLocalizationHelper.Jss15TT); // job types
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-            try
-            {
-                var stuff = this.df.ConvertJobSessSummaryToXml(scrub);
-
-                foreach (var stu in stuff)
-                {
-                    try
-                    {
-                        string t = string.Empty;
-                        t += "<tr>";
-
-                        t += this.form.TableDataLeftAligned(stu.JobName, VbrLocalizationHelper.Jss0);
-                        t += this.form.TableData(stu.ItemCount.ToString(), VbrLocalizationHelper.Jss1);
-                        t += this.form.TableData(stu.MinJobTime, VbrLocalizationHelper.Jss2);
-                        t += this.form.TableData(stu.MaxJobTime, VbrLocalizationHelper.Jss3);
-                        t += this.form.TableData(stu.AvgJobTime, VbrLocalizationHelper.Jss4);
-                        t += this.form.TableData(stu.SessionCount.ToString(), VbrLocalizationHelper.Jss5);
-                        t += this.form.TableData(stu.Fails.ToString(), "Fails");
-                        t += this.form.TableData(stu.Retries.ToString(), "Retries");
-                        t += this.form.TableData(stu.SuccessRate.ToString(), VbrLocalizationHelper.Jss6);
-                        t += this.form.TableData(stu.AvgBackupSize.ToString(), VbrLocalizationHelper.Jss7);
-                        t += this.form.TableData(stu.MaxBackupSize.ToString(), VbrLocalizationHelper.Jss8);
-                        t += this.form.TableData(stu.AvgDataSize.ToString(), VbrLocalizationHelper.Jss9);
-                        t += this.form.TableData(stu.MaxDataSize.ToString(), VbrLocalizationHelper.Jss10);
-                        t += this.form.TableData(stu.AvgChangeRate.ToString(), VbrLocalizationHelper.Jss11);
-                        t += this.form.TableData(stu.WaitCount.ToString(), VbrLocalizationHelper.Jss12);
-                        t += this.form.TableData(stu.MaxWait, VbrLocalizationHelper.Jss13);
-                        t += this.form.TableData(stu.AvgWait, VbrLocalizationHelper.Jss14);
-                        if (CGlobals.DEBUG)
-                        {
-                            this.log.Debug("Job Name = " + stu.JobName);
-                            this.log.Debug("Job Type = " + stu.JobType);
-                        }
-
-                        string jobType = CJobTypesParser.GetJobType(stu.JobType);
-                        t += this.form.TableData(jobType, VbrLocalizationHelper.Jss15);
-
-                        t += "</tr>";
-
-                        s += t;
-                    }
-                    catch (Exception ex)
-                    {
-                        this.log.Error("Job Session Summary Table failed to add row for job: " + stu.JobName);
-                        this.log.Error("\t" + ex.Message);
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Job Session Summary Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-            this.log.Info("Job Session Summary Table added");
-
-            // JSON job session summary
-            try
-            {
-                var stuff = this.df.ConvertJobSessSummaryToXml(scrub);
-                List<string> headers = new() { "JobName", "ItemCount", "MinJobTime", "MaxJobTime", "AvgJobTime", "SessionCount", "Fails", "Retries", "SuccessRate", "AvgBackupSize", "MaxBackupSize", "AvgDataSize", "MaxDataSize", "AvgChangeRate", "WaitCount", "MaxWait", "AvgWait", "JobTypes" };
-                List<List<string>> rows = stuff.Select(stu => new List<string>
-                {
-                    stu.JobName,
-                    stu.ItemCount.ToString(),
-                    stu.MinJobTime,
-                    stu.MaxJobTime,
-                    stu.AvgJobTime,
-                    stu.SessionCount.ToString(),
-                    stu.Fails.ToString(),
-                    stu.Retries.ToString(),
-                    stu.SuccessRate.ToString(),
-                    stu.AvgBackupSize.ToString(),
-                    stu.MaxBackupSize.ToString(),
-                    stu.AvgDataSize.ToString(),
-                    stu.MaxDataSize.ToString(),
-                    stu.AvgChangeRate.ToString(),
-                    stu.WaitCount.ToString(),
-                    stu.MaxWait,
-                    stu.AvgWait,
-                    CJobTypesParser.GetJobType(stu.JobType),
-                }).ToList();
-                SetSection("jobSessionSummary", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture jobSessionSummary JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CJobSessionSummaryTable();
+            return table.RenderFlat(scrub);
         }
 
         // Export the aggregated JSON report to a file
@@ -2558,803 +1820,22 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
 
         public string AddJobSessSummTableByJob(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("jobsesssum", VbrLocalizationHelper.JssTitle, VbrLocalizationHelper.JssBtn, CGlobals.ReportDays);
-            s += "</table>";
-
-            string summary = this.sum.JobInfo();
-
-            try
-            {
-                CCsvParser csvparser = new();
-                var source = csvparser.JobCsvParser().ToList();
-                source.OrderBy(x => x.Name);
-                var stuff = this.df.ConvertJobSessSummaryToXml(scrub);
-                var jobTypes = stuff.Select(x => x.JobType).Distinct().ToList();
-
-                List<CJobSummaryTypes> OffloadJobs = new();
-
-                // log.Debug("Sessions count = " + stuff.Count);
-                try
-                {
-                    foreach (var jType in jobTypes)
-                    {
-                        if (CGlobals.DEBUG)
-                        {
-                            this.log.Debug("Job Type = " + jType);
-                        }
-
-                        bool skipTotals = false;
-                        var jobType = jType;
-
-                        // change job type for "Totals" line of all jobs:
-                        if (jType == string.Empty || jType == null)
-                        {
-                            // log.Debug("Job Type is empty...");
-                            // jobType = "Summary of All";
-                        }
-
-                        var realType = CJobTypesParser.GetJobType(jobType);
-
-                        // log.Debug("Real Job Type = " + realType);
-                        string sectionHeader = realType;
-                        if (jType == null)
-                        {
-                            sectionHeader = "Summary of All";
-                        }
-
-                        string jobTable = this.form.SectionStartWithButton("jobTable", sectionHeader + " Jobs", string.Empty);
-                        s += jobTable;
-                        s += this.SetJobSessionsHeaders();
-                        var res = stuff.Where(x => x.JobType == jobType).ToList();
-
-                        // log.Debug("Jobs of type " + jType + " found:" + res.Count);
-
-                        // define variable to capture the TOTALS
-                        int totalItemsCount = 0;
-                        double totalSessionCount = 0;
-                        int totalFails = 0;
-                        int totalRetries = 0;
-                        double totalSuccessRate = 0;
-                        double totalAvgBackupSize = 0;
-                        double totalMaxBackupSize = 0;
-                        double totalAvgDataSize = 0;
-                        double totalMaxDataSize = 0;
-                        double totalAvgChangeRate = 0;
-                        int totalWaitCount = 0;
-
-                        foreach (var stu in res)
-                        {
-                            // if the job name contains "Offload" then add it to the OffloadJobs list
-                            if (stu.JobName.Contains("Offload"))
-                            {
-                                OffloadJobs.Add(stu);
-                                continue;
-                            }
-
-                            if (stu.JobName == "Total")
-                            {
-                                skipTotals = true;
-                            }
-
-                            // log.Debug("job name = " + stu.JobName);
-                            if (stu.JobType != jobType && stu.JobName != "Totals")
-                            {
-                                // log.Debug("Job Type: " + stu.JobType + " does not match " + jobType + "... skipping");
-                                continue;
-                            }
-
-                            try
-                            {
-                                // log.Debug("Adding row for job: " + stu.JobName);
-                                string t = string.Empty;
-                                t += "<tr>";
-
-                                t += this.form.TableDataLeftAligned(stu.JobName, VbrLocalizationHelper.Jss0);
-                                t += this.form.TableData(stu.ItemCount.ToString(), VbrLocalizationHelper.Jss1);
-                                t += this.form.TableData(stu.MinJobTime, VbrLocalizationHelper.Jss2);
-                                t += this.form.TableData(stu.MaxJobTime, VbrLocalizationHelper.Jss3);
-                                t += this.form.TableData(stu.AvgJobTime, VbrLocalizationHelper.Jss4);
-                                t += this.form.TableData(stu.SessionCount.ToString(), VbrLocalizationHelper.Jss5);
-                                t += this.form.TableData(stu.Fails.ToString(), "Fails");
-                                t += this.form.TableData(stu.Retries.ToString(), "Retries");
-                                t += this.form.TableData(stu.SuccessRate.ToString(), VbrLocalizationHelper.Jss6);
-                                t += this.form.TableData(stu.AvgBackupSize.ToString(), VbrLocalizationHelper.Jss7);
-                                t += this.form.TableData(stu.MaxBackupSize.ToString(), VbrLocalizationHelper.Jss8);
-                                t += this.form.TableData(stu.AvgDataSize.ToString(), VbrLocalizationHelper.Jss9);
-                                t += this.form.TableData(stu.MaxDataSize.ToString(), VbrLocalizationHelper.Jss10);
-                                t += this.form.TableData(stu.AvgChangeRate.ToString(), VbrLocalizationHelper.Jss11);
-                                t += this.form.TableData(stu.WaitCount.ToString(), VbrLocalizationHelper.Jss12);
-                                t += this.form.TableData(stu.MaxWait, VbrLocalizationHelper.Jss13);
-                                t += this.form.TableData(stu.AvgWait, VbrLocalizationHelper.Jss14);
-                                string jt = CJobTypesParser.GetJobType(stu.JobType);
-                                t += this.form.TableData(jt, VbrLocalizationHelper.Jss15);
-
-                                t += "</tr>";
-
-                                s += t;
-
-                                totalItemsCount += stu.ItemCount;
-                                totalSessionCount += stu.SessionCount;
-                                totalFails += stu.Fails;
-                                totalRetries += stu.Retries;
-
-                                // double successPercent = (totalSessionCount - (double)totalFails + totalRetries) / totalSessionCount * 100;
-                                // totalSuccessRate += (double)Math.Round(successPercent, 2);
-                                totalAvgBackupSize += stu.AvgBackupSize;
-                                totalMaxBackupSize += stu.MaxBackupSize;
-                                totalAvgDataSize += stu.AvgDataSize;
-                                totalMaxDataSize += stu.MaxDataSize;
-
-                                // totalAvgChangeRate += Math.Round(totalAvgDataSize / totalMaxDataSize * stu.AvgChangeRate, 2);
-                                totalWaitCount += stu.WaitCount;
-                            }
-
-                            catch (Exception ex)
-                            {
-                                this.log.Error("Job Session Summary Table failed to add row for job: " + stu.JobName);
-                                this.log.Error("\t" + ex.Message);
-                            }
-                        }
-
-                        // clean up totals:
-                        double successPercent = (totalSessionCount - (double)totalFails + totalRetries) / totalSessionCount * 100;
-                        totalSuccessRate = (double)Math.Round(successPercent, 2);
-                        if (totalAvgDataSize == 0 && totalMaxDataSize == 0)
-                        {
-                            totalAvgChangeRate = 0;
-                        }
-                        else
-                        {
-                            totalAvgChangeRate = Math.Round(totalAvgDataSize / totalMaxDataSize * 100, 2);
-                        }
-
-                        // log.Debug("Total avg change rate = " + totalAvgChangeRate);
-                        // add totals line:
-                        if (!skipTotals)
-                        {
-                            string totalRow = string.Empty;
-                            totalRow += "<tr>";
-                            totalRow += this.form.TableDataLeftAligned("TOTALS", string.Empty);
-                            totalRow += this.form.TableData(totalItemsCount.ToString(), string.Empty);
-                            totalRow += this.form.TableData(string.Empty, string.Empty);
-                            totalRow += this.form.TableData(string.Empty, string.Empty);
-                            totalRow += this.form.TableData(string.Empty, string.Empty);
-                            totalRow += this.form.TableData(totalSessionCount.ToString(), string.Empty);
-                            totalRow += this.form.TableData(totalFails.ToString(), string.Empty);
-                            totalRow += this.form.TableData(totalRetries.ToString(), string.Empty);
-                            totalRow += this.form.TableData(totalSuccessRate.ToString(), string.Empty);
-                            totalRow += this.form.TableData(Math.Round(totalAvgBackupSize, 2).ToString(), string.Empty);
-                            totalRow += this.form.TableData(Math.Round(totalMaxBackupSize, 2).ToString(), string.Empty);
-                            totalRow += this.form.TableData(Math.Round(totalAvgDataSize, 2).ToString(), string.Empty);
-                            totalRow += this.form.TableData(Math.Round(totalMaxDataSize, 2).ToString(), string.Empty);
-                            if (totalAvgChangeRate == double.NaN)
-                            {
-                                totalAvgChangeRate = 0;
-                            }
-
-                            totalRow += this.form.TableData(totalAvgChangeRate.ToString(), string.Empty);
-                            totalRow += this.form.TableData(totalWaitCount.ToString(), string.Empty);
-                            totalRow += this.form.TableData(string.Empty, string.Empty);
-                            totalRow += this.form.TableData(string.Empty, string.Empty);
-                            totalRow += this.form.TableData(string.Empty, string.Empty);
-                            s += totalRow;
-                        }
-
-                        // table summary/totals
-
-                        // end each table/section
-                        s += this.form.SectionEnd(summary);
-                    }
-
-                    s += this.AddOffloadsTable(OffloadJobs);
-                }
-                catch (Exception e)
-                {
-                    this.log.Error("Job Info Data import failed. ERROR:");
-                    this.log.Error("\t" + e.Message);
-                }
-
-                // end of FE up one line...
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Jobs Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON job session summary by job
-            try
-            {
-                var stuff = this.df.ConvertJobSessSummaryToXml(scrub);
-                var ordered = stuff.OrderBy(stu => stu.JobName).ToList();
-                List<string> headers = new() { "JobName", "ItemCount", "MinJobTime", "MaxJobTime", "AvgJobTime", "SessionCount", "Fails", "Retries", "SuccessRate", "AvgBackupSize", "MaxBackupSize", "AvgDataSize", "MaxDataSize", "AvgChangeRate", "WaitCount", "MaxWait", "AvgWait", "JobTypes" };
-                List<List<string>> rows = ordered.Select(stu => new List<string>
-                {
-                    stu.JobName,
-                    stu.ItemCount.ToString(),
-                    stu.MinJobTime,
-                    stu.MaxJobTime,
-                    stu.AvgJobTime,
-                    stu.SessionCount.ToString(),
-                    stu.Fails.ToString(),
-                    stu.Retries.ToString(),
-                    stu.SuccessRate.ToString(),
-                    stu.AvgBackupSize.ToString(),
-                    stu.MaxBackupSize.ToString(),
-                    stu.AvgDataSize.ToString(),
-                    stu.MaxDataSize.ToString(),
-                    stu.AvgChangeRate.ToString(),
-                    stu.WaitCount.ToString(),
-                    stu.MaxWait,
-                    stu.AvgWait,
-                    CJobTypesParser.GetJobType(stu.JobType),
-                }).ToList();
-                SetSection("jobSessionSummaryByJob", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture jobSessionSummaryByJob JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CJobSessionSummaryTable();
+            return table.RenderByJob(scrub);
         }
 
-        private string AddOffloadsTable(List<CJobSummaryTypes> offloadJobs)
-        {
-            string s = string.Empty;
-            try
-            {
-                // log.Debug("Checking Job Type: " + jType);
-                // var translatedJobType = CJobTypesParser.GetJobType(jType);
-                // log.Debug("\tTranslated Job Type: " + translatedJobType);
-                var realType = "Offload";
-                string jobTable = this.form.SectionStartWithButton("jobTable", realType + " Jobs", string.Empty);
-                s += jobTable;
-                s += this.SetJobSessionsHeaders();
-
-                // var res = stuff.Where(x => x.JobType == jType).ToList();
-                // log.Debug("Jobs of type " + jType + " found:" + res.Count);
-
-                // var res2 = stuff.Where(x => x.JobType == realType).ToList();
-
-                // log.Debug("Jobs of type " + jType + " found:" + res2.Count);
-                // define variable to capture the TOTALS
-                int totalItemsCount = 0;
-                double totalSessionCount = 0;
-                int totalFails = 0;
-                int totalRetries = 0;
-                double totalSuccessRate = 0;
-                double totalAvgBackupSize = 0;
-                double totalMaxBackupSize = 0;
-                double totalAvgDataSize = 0;
-                double totalMaxDataSize = 0;
-                double totalAvgChangeRate = 0;
-                int totalWaitCount = 0;
-
-                foreach (var stu in offloadJobs)
-                {
-                    // if the job name contains "Offload" then add it to the OffloadJobs list
-
-                    // log.Debug("job type of individual session = " + stu.JobType);
-
-                    // if (stu.JobType != jType)
-                    // {
-                    //     log.Debug("Job Type: " + stu.JobType + " does not match " + jType + "... skipping");
-                    //     continue;
-                    // }
-                    try
-                    {
-                        // log.Debug("Adding row for job: " + stu.JobName);
-                        string t = string.Empty;
-                        t += "<tr>";
-
-                        t += this.form.TableDataLeftAligned(stu.JobName, VbrLocalizationHelper.Jss0);
-                        t += this.form.TableData(stu.ItemCount.ToString(), VbrLocalizationHelper.Jss1);
-                        t += this.form.TableData(stu.MinJobTime, VbrLocalizationHelper.Jss2);
-                        t += this.form.TableData(stu.MaxJobTime, VbrLocalizationHelper.Jss3);
-                        t += this.form.TableData(stu.AvgJobTime, VbrLocalizationHelper.Jss4);
-                        t += this.form.TableData(stu.SessionCount.ToString(), VbrLocalizationHelper.Jss5);
-                        t += this.form.TableData(stu.Fails.ToString(), "Fails");
-                        t += this.form.TableData(stu.Retries.ToString(), "Retries");
-                        t += this.form.TableData(stu.SuccessRate.ToString(), VbrLocalizationHelper.Jss6);
-                        t += this.form.TableData(stu.AvgBackupSize.ToString(), VbrLocalizationHelper.Jss7);
-                        t += this.form.TableData(stu.MaxBackupSize.ToString(), VbrLocalizationHelper.Jss8);
-                        t += this.form.TableData(stu.AvgDataSize.ToString(), VbrLocalizationHelper.Jss9);
-                        t += this.form.TableData(stu.MaxDataSize.ToString(), VbrLocalizationHelper.Jss10);
-                        t += this.form.TableData(stu.AvgChangeRate.ToString(), VbrLocalizationHelper.Jss11);
-                        t += this.form.TableData(stu.WaitCount.ToString(), VbrLocalizationHelper.Jss12);
-                        t += this.form.TableData(stu.MaxWait, VbrLocalizationHelper.Jss13);
-                        t += this.form.TableData(stu.AvgWait, VbrLocalizationHelper.Jss14);
-                        string jobType = CJobTypesParser.GetJobType(stu.JobType);
-                        t += this.form.TableData(jobType, VbrLocalizationHelper.Jss15);
-
-                        t += "</tr>";
-
-                        s += t;
-
-                        totalItemsCount += stu.ItemCount;
-                        totalSessionCount += stu.SessionCount;
-                        totalFails += stu.Fails;
-                        totalRetries += stu.Retries;
-
-                        // double successPercent = (totalSessionCount - (double)totalFails + totalRetries) / totalSessionCount * 100;
-                        // totalSuccessRate += (double)Math.Round(successPercent, 2);
-                        totalAvgBackupSize += stu.AvgBackupSize;
-                        totalMaxBackupSize += stu.MaxBackupSize;
-                        totalAvgDataSize += stu.AvgDataSize;
-                        totalMaxDataSize += stu.MaxDataSize;
-
-                        // totalAvgChangeRate += Math.Round(totalAvgDataSize / totalMaxDataSize * stu.AvgChangeRate, 2);
-                        totalWaitCount += stu.WaitCount;
-                    }
-
-                    catch (Exception ex)
-                    {
-                        this.log.Error("Job Session Summary Table failed to add row for job: " + stu.JobName);
-                        this.log.Error("\t" + ex.Message);
-                    }
-                }
-
-                // clean up totals:
-                double successPercent = (totalSessionCount - (double)totalFails + totalRetries) / totalSessionCount * 100;
-                totalSuccessRate = (double)Math.Round(successPercent, 2);
-
-                totalAvgChangeRate = Math.Round(totalAvgDataSize / totalMaxDataSize * 100, 2);
-
-                // add totals line:
-                string totalRow = string.Empty;
-                totalRow += "<tr>";
-                totalRow += this.form.TableDataLeftAligned("TOTALS", string.Empty);
-                totalRow += this.form.TableData(totalItemsCount.ToString(), string.Empty);
-                totalRow += this.form.TableData(string.Empty, string.Empty);
-                totalRow += this.form.TableData(string.Empty, string.Empty);
-                totalRow += this.form.TableData(string.Empty, string.Empty);
-                totalRow += this.form.TableData(totalSessionCount.ToString(), string.Empty);
-                totalRow += this.form.TableData(totalFails.ToString(), string.Empty);
-                totalRow += this.form.TableData(totalRetries.ToString(), string.Empty);
-                totalRow += this.form.TableData(totalSuccessRate.ToString(), string.Empty);
-                totalRow += this.form.TableData(Math.Round(totalAvgBackupSize, 2).ToString(), string.Empty);
-                totalRow += this.form.TableData(Math.Round(totalMaxBackupSize, 2).ToString(), string.Empty);
-                totalRow += this.form.TableData(Math.Round(totalAvgDataSize, 2).ToString(), string.Empty);
-                totalRow += this.form.TableData(Math.Round(totalMaxDataSize, 2).ToString(), string.Empty);
-                totalRow += this.form.TableData(totalAvgChangeRate.ToString(), string.Empty);
-                totalRow += this.form.TableData(totalWaitCount.ToString(), string.Empty);
-                totalRow += this.form.TableData(string.Empty, string.Empty);
-                totalRow += this.form.TableData(string.Empty, string.Empty);
-                totalRow += this.form.TableData(string.Empty, string.Empty);
-                s += totalRow;
-
-                // table summary/totals
-
-                // end each table/section
-                s += this.form.SectionEnd(string.Empty);
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Job Info Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            return s;
-        }
 
         public string AddJobInfoTable(bool scrub)
         {
-            string s = this.form.SectionStartWithButton("jobs", VbrLocalizationHelper.JobInfoTitle, VbrLocalizationHelper.JobInfoBtn);
-            s += "</table>";
-
-            string summary = this.sum.JobInfo();
-
-            try
-            {
-                CCsvParser csvparser = new();
-                var source = csvparser.JobCsvParser().ToList();
-                source.OrderBy(x => x.Name);
-                var jobTypes = source.Select(x => x.JobType).Distinct().ToList();
-
-                try
-                {
-                    foreach (var jType in jobTypes)
-                    {
-                        double tSizeGB = 0;
-                        double onDiskTotalGB = 0;
-
-                        bool useSourceSize = !(jType == "NasBackupCopy" || jType == "Copy");
-
-                        var realType = CJobTypesParser.GetJobType(jType);
-                        string jobTable = this.form.SectionStartWithButton("jobTable", realType + " Jobs", string.Empty);
-                        s += jobTable;
-                        s += this.SetGenericJobTablHeader(useSourceSize, jType);
-                        var res = source.Where(x => x.JobType == jType).ToList();
-                        foreach (var job in res)
-                        {
-                            // object x = null;
-                            double onDiskGB = 0;
-                            double sourceSizeGB = 0;
-
-                            if (job.JobType != jType)
-                            {
-                                continue;
-                            }
-
-                            string row = string.Empty;
-                            if (jType == "NasBackup")
-                            {
-                                var x = csvparser.GetDynamicNasBackup().ToList();
-
-                                var diskGb = x.Where(x => x.name == job.Name)
-                                    .Select(x => x.ondiskgb)
-                                    .FirstOrDefault();
-                                double.TryParse(diskGb, out onDiskGB);
-                                onDiskGB = Math.Round(onDiskGB, 2);
-                                onDiskTotalGB += onDiskGB;
-
-                                var sourceGb = x.Where(x => x.name == job.Name)
-                                    .Select(x => x.sourcegb)
-                                    .FirstOrDefault();
-                                double.TryParse(sourceGb, out sourceSizeGB);
-                                sourceSizeGB = Math.Round(sourceSizeGB, 2);
-                            }
-                            else
-                            {
-                                //double.TryParse(job.OnDiskGB, out onDiskGB);
-                                onDiskGB = Math.Round(job.OnDiskGB ?? 0, 2);
-
-                            }
-
-                            string jobName = job.Name;
-                            string repoName = job.RepoName;
-                            if (scrub)
-                            {
-                                jobName = CGlobals.Scrubber.ScrubItem(jobName, ScrubItemType.Job);
-                                repoName = CGlobals.Scrubber.ScrubItem(repoName, ScrubItemType.Repository);
-                            }
-
-                            row += "<tr>";
-                            row += this.form.TableDataLeftAligned(jobName, string.Empty);
-                            row += this.form.TableData(repoName, string.Empty);
-
-                            if (useSourceSize)
-                            {
-                                if (jType == "NasBackup")
-                                {
-                                    row += this.form.TableData(sourceSizeGB.ToString(), string.Empty);
-                                    tSizeGB += sourceSizeGB;
-                                }
-                                else
-                                {
-                                    double trueSizeGB = Math.Round(job.OriginalSize / 1024 / 1024 / 1024, 2);
-                                    double trueSizeTB = Math.Round(job.OriginalSize / 1024 / 1024 / 1024 / 1024, 2);
-                                    double trueSizeMB = Math.Round(job.OriginalSize / 1024 / 1024, 2);
-                                    tSizeGB += trueSizeGB;
-                                    if (trueSizeGB > 999)
-                                    {
-                                        row += this.form.TableData(trueSizeTB.ToString() + " TB", string.Empty);
-                                    }
-                                    else if (trueSizeGB < 1)
-                                    {
-                                        row += this.form.TableData(trueSizeMB.ToString() + " MB", string.Empty);
-                                    }
-                                    else
-                                    {
-                                        row += this.form.TableData(trueSizeGB.ToString() + " GB", string.Empty);
-                                    }
-                                }
-                            }
-                            else
-                            {
-                            }
-
-                            row += this.form.TableData(onDiskGB.ToString(), string.Empty);
-
-                            // row+= _form.TableData(trueSizeGB.ToString() + " GB", "");
-                            // row+= _form.TableData(job.RetentionType, "");
-                            row += job.RetentionType == "Cycles" ? this.form.TableData("Points", string.Empty) : this.form.TableData(job.RetentionType, string.Empty);
-                            string retentionValue = job.RetentionType == "Cycles"
-                                ? job.RetentionCount
-                                : job.RetainDaysToKeep;
-                            row += this.form.TableData(retentionValue, string.Empty);
-
-                            // row += _form.TableData(job.StgEncryptionEnabled, "");
-                            row += job.StgEncryptionEnabled == "True" ? this.form.TableData(this.form.True, string.Empty) : this.form.TableData(this.form.False, string.Empty);
-                            var jobType = CJobTypesParser.GetJobType(job.JobType);
-                            row += this.form.TableData(jobType, string.Empty);
-
-                            // row += _form.TableData("", "");
-                            string compressionLevel = string.Empty;
-                            if (job.CompressionLevel == "9")
-                            {
-                                compressionLevel = "Extreme";
-                            }
-                            else if (job.CompressionLevel == "6")
-                            {
-                                compressionLevel = "High";
-                            }
-                            else if (job.CompressionLevel == "5")
-                            {
-                                compressionLevel = "Optimal";
-                            }
-                            else if (job.CompressionLevel == "4")
-                            {
-                                compressionLevel = "Dedupe-Friendly";
-                            }
-                            else if (job.CompressionLevel == "0")
-                            {
-                                compressionLevel = "None";
-                            }
-
-                            row += this.form.TableData(compressionLevel, string.Empty);
-
-                            string blockSize = string.Empty;
-                            if (job.BlockSize == "KbBlockSize1024")
-                            {
-                                blockSize = "1 MB";
-                            }
-                            else if (job.BlockSize == "KbBlockSize512")
-                            {
-                                blockSize = "512 KB";
-                            }
-                            else if (job.BlockSize == "KbBlockSize256")
-                            {
-                                blockSize = "256 KB";
-                            }
-                            else if (job.BlockSize == "KbBlockSize4096")
-                            {
-                                blockSize = "4 MB";
-                            }
-                            else if (job.BlockSize == "KbBlockSize8192")
-                            {
-                                blockSize = "8 MB";
-                            }
-
-                            row += this.form.TableData(blockSize, string.Empty);
-                            if (job.GfsMonthlyEnabled || job.GfsWeeklyIsEnabled || job.GfsYearlyEnabled)
-                            {
-                                row += this.form.TableData(this.form.True, string.Empty);
-                                var gfsParts = new List<string>();
-                                if (job.GfsWeeklyIsEnabled)
-                                    gfsParts.Add("Weekly: " + job.GfsWeeklyCount);
-                                if (job.GfsMonthlyEnabled)
-                                    gfsParts.Add("Monthly: " + job.GfsMonthlyCount);
-                                if (job.GfsYearlyEnabled)
-                                    gfsParts.Add("Yearly: " + job.GfsYearlyCount);
-                                string GfsString = string.Join("<br> ", gfsParts);
-                                row += this.form.TableData(GfsString, string.Empty);
-                            }
-                            else
-                            {
-                                row += this.form.TableData(this.form.False, string.Empty);
-                                row += this.form.TableData(string.Empty, string.Empty);
-                            }
-
-                            row += job.EnableFullBackup ? this.form.TableData(this.form.True, string.Empty) : this.form.TableData(this.form.False, string.Empty);
-                            try
-                            {
-                                if (job.Algorithm == "Increment" && job.TransformFullToSyntethic == true)
-                                {
-                                    row += this.form.TableData(this.form.True, string.Empty);
-                                }
-                                else
-                                {
-                                    row += this.form.TableData(this.form.False, string.Empty);
-                                }
-                            }
-                            catch
-                            {
-                                row += this.form.TableData(this.form.False, string.Empty);
-                            }
-
-                            row += job.Algorithm == "Syntethic" ? this.form.TableData("Reverse Incremental", string.Empty) : this.form.TableData("Forward Incremental", string.Empty);
-
-                            row += job.IndexingType != "None" ? this.form.TableData(this.form.True, string.Empty) : this.form.TableData(this.form.False, string.Empty);
-
-                            // AAIP/Guest Processing columns - only for job types that support it
-                            if (JobTypeSupportsAaip(jType))
-                            {
-                                row += job.AAIPEnabled == "True" ? this.form.TableData(this.form.True, string.Empty) : this.form.TableData(this.form.False, string.Empty);
-                                row += job.VSSIgnoreErrors == "True" ? this.form.TableData(this.form.True, string.Empty) : this.form.TableData(this.form.False, string.Empty);
-                                row += job.GuestFSIndexingEnabled == "True" ? this.form.TableData(this.form.True, string.Empty) : this.form.TableData(this.form.False, string.Empty);
-                            }
-
-                            row += "</tr>";
-
-                            s += row;
-                        }
-
-                        // table summary/totals
-                        if (useSourceSize)
-                        {
-                            s += "<tr>";
-                            s += this.form.TableDataLeftAligned("Totals", string.Empty);
-                            s += this.form.TableData(string.Empty, string.Empty);
-
-                            // double totalSizeGB = Math.Round(tSizeGB / 1024 / 1024 / 1024, 2);
-                            double totalSizeTB = Math.Round(tSizeGB / 1024, 2);
-                            double totalSizeMB = Math.Round(tSizeGB * 1024, 2);
-
-                            double diskTotalTB = Math.Round(onDiskTotalGB / 1024, 2);
-                            double diskTotalMB = Math.Round(onDiskTotalGB * 1024, 2);
-                            if (tSizeGB > 999)
-                            {
-                                s += this.form.TableData(totalSizeTB.ToString() + " TB", string.Empty);
-                            }
-                            else if (tSizeGB < 1)
-                            {
-                                s += this.form.TableData(totalSizeMB.ToString() + " MB", string.Empty);
-                            }
-                            else
-                            {
-                                s += this.form.TableData(tSizeGB.ToString() + " GB", string.Empty);
-                            }
-
-                            if (diskTotalTB > 1)
-                            {
-                                s += this.form.TableData(onDiskTotalGB.ToString(), "TB");
-                            }
-                            else if (onDiskTotalGB > 1)
-                            {
-                                s += this.form.TableData(onDiskTotalGB.ToString(), "GB");
-                            }
-                            else
-                            {
-                                s += this.form.TableData(onDiskTotalGB.ToString(), "MB");
-                            }
-                        }
-
-                        // end each table/section
-                        s += this.form.SectionEnd(summary);
-                    }
-
-                    // add tape table
-                    try
-                    {
-                        CTapeJobInfoTable tapeTable = new();
-                        string tt = tapeTable.TapeJobTable();
-                        if (tt != string.Empty)
-                        {
-                            string tableButton = this.form.SectionStartWithButton("jobTable", "Tape Jobs", string.Empty);
-                            s += tableButton;
-                            s += tt;
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        this.log.Error("Tape Job Data import failed. ERROR:");
-                        this.log.Error("\t" + e.Message);
-                    }
-
-                    // Add Entra Table
-                    try
-                    {
-                        CEntraJobsTable entraTable = new();
-                        string et = entraTable.Table();
-                        if (et != null && et.Length > 0)
-                        {
-                            // debug
-                            this.log.Debug("Entra jobs table length: " + et.Length);
-                            string tableButton = this.form.SectionStartWithButton("jobTable", "Entra Jobs", string.Empty);
-                            s += tableButton;
-                            s += et;
-                            s += this.form.SectionEnd(summary);
-                        }
-                        else
-                        {
-                            this.log.Info("No Entra backup jobs detected - skipping Entra jobs table", false);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        this.log.Error("Entra Job Data import failed. ERROR:");
-                        this.log.Error("\t" + e.Message);
-                    }
-
-                    s += this.form.SectionEnd(summary);
-                }
-                catch (Exception e)
-                {
-                    this.log.Error("Job Info Data import failed. ERROR:");
-                    this.log.Error("\t" + e.Message);
-                }
-
-                // end of FE up one line...
-            }
-            catch (Exception e)
-            {
-                this.log.Error("Jobs Data import failed. ERROR:");
-                this.log.Error("\t" + e.Message);
-            }
-
-            s += this.form.SectionEnd(summary);
-
-            // JSON job info capture
-            try
-            {
-                CCsvParser csvparser = new();
-                var source = csvparser.JobCsvParser().ToList();
-                List<string> headers = new() { "JobName", "RepoName", "SourceSizeGB", "OnDiskGB", "RetentionScheme", "RetainDays", "Encrypted", "JobType", "CompressionLevel", "BlockSize", "GfsEnabled", "GfsDetails", "ActiveFullEnabled", "SyntheticFullEnabled", "BackupChainType", "IndexingEnabled", "AAIPEnabled", "VSSEnabled", "VSSIgnoreErrors", "GuestFSIndexing" };
-                List<List<string>> rows = new();
-
-                foreach (var job in source)
-                {
-                    string jobName = scrub ? CGlobals.Scrubber.ScrubItem(job.Name, ScrubItemType.Job) : job.Name;
-                    string repoName = scrub ? CGlobals.Scrubber.ScrubItem(job.RepoName, ScrubItemType.Repository) : job.RepoName;
-
-                    double sourceSizeGB = Math.Round(job.OriginalSize / 1024.0 / 1024.0 / 1024.0, 2);
-                    string compressionLevel = job.CompressionLevel switch
-                    {
-                        "9" => "Extreme",
-                        "6" => "High",
-                        "5" => "Optimal",
-                        "4" => "Dedupe-Friendly",
-                        "0" => "None",
-                        _ => job.CompressionLevel,
-                    };
-
-                    string blockSize = job.BlockSize switch
-                    {
-                        "KbBlockSize1024" => "1 MB",
-                        "KbBlockSize512" => "512 KB",
-                        "KbBlockSize256" => "256 KB",
-                        "KbBlockSize4096" => "4 MB",
-                        "KbBlockSize8192" => "8 MB",
-                        _ => job.BlockSize,
-                    };
-
-                    bool gfsEnabled = job.GfsMonthlyEnabled || job.GfsWeeklyIsEnabled || job.GfsYearlyEnabled;
-                    var gfsDetailParts = new List<string>();
-                    if (job.GfsWeeklyIsEnabled)
-                        gfsDetailParts.Add($"Weekly:{job.GfsWeeklyCount}");
-                    if (job.GfsMonthlyEnabled)
-                        gfsDetailParts.Add($"Monthly:{job.GfsMonthlyCount}");
-                    if (job.GfsYearlyEnabled)
-                        gfsDetailParts.Add($"Yearly:{job.GfsYearlyCount}");
-                    string gfsDetails = gfsEnabled ? string.Join(",", gfsDetailParts) : string.Empty;
-                    bool syntheticFull = job.Algorithm == "Increment" && job.TransformFullToSyntethic;
-                    string backupChainType = job.Algorithm == "Syntethic" ? "Reverse Incremental" : "Forward Incremental";
-                    bool indexingEnabled = job.IndexingType != "None";
-
-                    //debug log for ondiskgb
-                    this.log.Debug($"Job: {jobName}, OnDiskGB: {job.OnDiskGB}");
-
-                    string retentionValue = job.RetentionType == "Cycles" ? job.RetentionCount : job.RetainDaysToKeep;
-                    rows.Add(new List<string>
-                    {
-                        jobName,
-                        repoName,
-                        sourceSizeGB.ToString(),
-                        (job.OnDiskGB ?? 0).ToString(), //"0", // OnDisk calculated separately for NAS
-                        job.RetentionType == "Cycles" ? "Points" : job.RetentionType,
-                        retentionValue,
-                        job.StgEncryptionEnabled,
-                        CJobTypesParser.GetJobType(job.JobType),
-                        compressionLevel,
-                        blockSize,
-                        gfsEnabled.ToString(),
-                        gfsDetails,
-                        job.EnableFullBackup.ToString(),
-                        syntheticFull.ToString(),
-                        backupChainType,
-                        indexingEnabled.ToString(),
-                        job.AAIPEnabled?.ToString() ?? "",
-                        job.VSSEnabled?.ToString() ?? "",
-                        job.VSSIgnoreErrors?.ToString() ?? "",
-                        job.GuestFSIndexingEnabled?.ToString() ?? "",
-                    });
-                }
-
-                SetSection("jobInfo", headers, rows, summary);
-            }
-            catch (Exception ex)
-            {
-                this.log.Error("Failed to capture jobInfo JSON section: " + ex.Message);
-            }
-
-            return s;
+            var table = new CJobInfoTable();
+            return table.Render(scrub);
+        }
+        public void AddSessionsFiles(bool scrub)
+        {
+            this.df.JobSessionInfoToXml(scrub);
         }
 
-        // --- Requirements tables (new) ---
+        // --- Requirements tables ---
         public string AddServersRequirementsTable(bool scrub)
         {
             return this.AddRequirementsTable(
@@ -3405,20 +1886,15 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
                 {
                     foreach (var r in rows)
                     {
-                        // expected indices:
-                        // 0 Server, 1 Type, 2 ReqCores, 3 AvailCores, 4 ReqRam, 5 AvailRam, 6 ConTasks, 7 SugTasks, 8 Names
                         string server = r[0];
                         string names = r[8];
 
                         if (doScrub)
                         {
                             server = this.scrub.ScrubItem(server, ScrubItemType.Server);
-
-                            // "Names" might contain server/proxy/repo names; scrubging as Server is usually "good enough"
                             names = this.scrub.ScrubItem(names, ScrubItemType.Server);
                         }
 
-                        // Safely parse numeric fields and highlight cells where requirements exceed availability
                         double reqCoresVal = 0, availCoresVal = 0, reqRamVal = 0, availRamVal = 0, concurrentTasksVal = 0, suggestedTasksVal = 0;
                         bool parsedReqCores = r.Length > 2 && double.TryParse(r[2], NumberStyles.Any, CultureInfo.InvariantCulture, out reqCoresVal);
                         bool parsedAvailCores = r.Length > 3 && double.TryParse(r[3], NumberStyles.Any, CultureInfo.InvariantCulture, out availCoresVal);
@@ -3435,7 +1911,6 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
                         s += this.form.TableData(server, string.Empty);
                         s += this.form.TableData(r[1], string.Empty);
 
-                        // Required Cores (highlight if required > available)
                         if (reqCoresWarn)
                         {
                             s += this.form.TableData(r[2], string.Empty, 1);
@@ -3445,10 +1920,8 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
                             s += this.form.TableData(r[2], string.Empty);
                         }
 
-                        // Available Cores
                         s += this.form.TableData(r[3], string.Empty);
 
-                        // Required RAM (highlight if required > available)
                         if (reqRamWarn)
                         {
                             s += this.form.TableData(r[4], string.Empty, 1);
@@ -3458,10 +1931,8 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
                             s += this.form.TableData(r[4], string.Empty);
                         }
 
-                        // Available RAM
                         s += this.form.TableData(r[5], string.Empty);
 
-                        // Concurrent Tasks (highlight if concurrent > suggested)
                         if (conTasksWarn)
                         {
                             s += this.form.TableData(r[6], string.Empty, 1);
@@ -3471,10 +1942,8 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
                             s += this.form.TableData(r[6], string.Empty);
                         }
 
-                        // Suggested Tasks
                         s += this.form.TableData(r[7], string.Empty);
 
-                        // Format names one per line
                         string formattedNames = names?.Replace("/ ", "<br>").Replace("/", "<br>") ?? "";
                         s += this.form.TableData(formattedNames, string.Empty);
                         s += "</tr>";
@@ -3492,7 +1961,7 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
 
             s += this.form.SectionEnd(summary);
 
-            // JSON capture (optional, but consistent with your file)
+            // JSON capture
             try
             {
                 List<string> headers = new()
@@ -3512,107 +1981,6 @@ this.form.TableHeader(VbrLocalizationHelper.SbrExt15, VbrLocalizationHelper.SbrE
             }
 
             return s;
-        }
-
-        private string SetJobSessionsHeaders()
-        {
-            string s = string.Empty;
-            s += this.form.TableHeaderLeftAligned(VbrLocalizationHelper.Jss0, VbrLocalizationHelper.Jss0TT); // job name
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss1, VbrLocalizationHelper.Jss1TT);// items
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss2, VbrLocalizationHelper.Jss2TT); // min time
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss3, VbrLocalizationHelper.Jss3TT);// max time
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss4, VbrLocalizationHelper.Jss4TT);// avg time
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss5, VbrLocalizationHelper.Jss5TT); // total sessions
-            s += this.form.TableHeader("Fails", "Total times job failed"); // fails
-            s += this.form.TableHeader("Retries", "Total times job retried");// retries
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss6, VbrLocalizationHelper.Jss6TT);// success rate
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss7, VbrLocalizationHelper.Jss7TT); // avg backup size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss8, VbrLocalizationHelper.Jss8TT);// max backup size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss9, VbrLocalizationHelper.Jss9TT); // avg data size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss10, "Used size of all objects in job."); // max data size
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss11, "Avg Data Size divided by Max Data Size (average processed data divided by total consumed size of all VMs in the job)"); // avg change rate
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss12, VbrLocalizationHelper.Jss12TT); // wait for res count
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss13, "Max Waits (dd:hh:mm:ss)"); // max wait
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss14, "Avg Waits (dd:hh:mm:ss)");// avg wait
-            s += this.form.TableHeader(VbrLocalizationHelper.Jss15, VbrLocalizationHelper.Jss15TT); // job types
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-
-            return s;
-        }
-
-        // Job types that support Application Aware Image Processing (guest OS quiescence)
-        private static readonly HashSet<string> AaipSupportedJobTypes = new(StringComparer.OrdinalIgnoreCase)
-        {
-            "Backup",           // VM backups (VMware/Hyper-V)
-            "EpAgentBackup",    // Physical/Endpoint backups
-            "Replica"           // VM replication
-        };
-
-        private static bool JobTypeSupportsAaip(string jobType) => AaipSupportedJobTypes.Contains(jobType);
-
-        private string SetGenericJobTablHeader(bool useSourceSize, string jobType = null)
-        {
-            bool showAaipColumns = jobType != null && JobTypeSupportsAaip(jobType);
-
-            string s = string.Empty;
-            s += this.form.TableHeaderLeftAligned(VbrLocalizationHelper.JobInfo0, VbrLocalizationHelper.JobInfo0TT); // Name
-            s += this.form.TableHeader(VbrLocalizationHelper.JobInfo1, VbrLocalizationHelper.JobInfo1TT); // Repo
-            if (useSourceSize)
-            {
-                s += this.form.TableHeader(VbrLocalizationHelper.JobInfo2, VbrLocalizationHelper.JobInfo2TT); // Source Size (GB)
-            }
-
-            s += this.form.TableHeader("Est. On Disk GB", "Estimated size of the backup data on-disk.");
-            s += this.form.TableHeader("Retention Scheme", "Is the job set to keep backups for X number of Days or Points");
-            s += this.form.TableHeader(VbrLocalizationHelper.JobInfo3, VbrLocalizationHelper.JobInfo3TT); // Restore Point Target
-            s += this.form.TableHeader(VbrLocalizationHelper.JobInfo4, VbrLocalizationHelper.JobInfo4TT); // Encrypted
-            s += this.form.TableHeader(VbrLocalizationHelper.JobInfo5, VbrLocalizationHelper.JobInfo5TT); // Job Type
-
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo6, VbrLocalizationHelper.JobInfo6TT); // Algorithm
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo7, VbrLocalizationHelper.JobInfo7TT); // Scheudle Enabled Time
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo8, VbrLocalizationHelper.JobInfo8TT); // Full Backup Days
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo9, VbrLocalizationHelper.JobInfo9TT); // Full Backup Schedule
-            ////s += _form.TableHeader(ResourceHandler.JobInfo10, ResourceHandler.JobInfo10TT);
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo11, VbrLocalizationHelper.JobInfo11TT); // transform full to synth
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo12, VbrLocalizationHelper.JobInfo12TT); // thransform inc to synth
-            // s += _form.TableHeader(VbrLocalizationHelper.JobInfo13, VbrLocalizationHelper.JobInfo13TT); // transform days
-
-            // New Table Area:
-            // s += _form.TableHeader("", "");
-            s += this.form.TableHeader("Compression Level", "Level of compression used in the job");
-            s += this.form.TableHeader("Block Size", "Block Size set for the job");
-            s += this.form.TableHeader("GFS Enabled", "True if any GFS Periods are enabled");
-            s += this.form.TableHeader("GFS Retention", "Details about the GFS Retention period");
-            s += this.form.TableHeader("Active Full Enabled", string.Empty);
-            s += this.form.TableHeader("Synthetic Full Enabled", string.Empty);
-            s += this.form.TableHeader("Backup Chain Type", "Type of backup chain used in the job");
-            s += this.form.TableHeader("Indexing Enabled", string.Empty);
-
-            // AAIP columns only for job types that support guest processing
-            if (showAaipColumns)
-            {
-                s += this.form.TableHeader("AAIP Enabled", "Enable application-aware processing checkbox in job settings");
-                s += this.form.TableHeader("VSS Ignore Errors", "Continue backup if VSS errors occur");
-                s += this.form.TableHeader("Guest FS Indexing", "Guest file system indexing enabled for file-level restore");
-            }
-
-            // s += _form.TableData("Totals", "");
-            // s += _form.TableHeader("", "");
-            // s += _form.TableHeader("", "");
-            // s += _form.TableHeader("", "");
-            // s += _form.TableHeader("", "");
-            // s += _form.TableHeader("", "");
-            // s += _form.TableHeader("", "");
-            s += this.form.TableHeaderEnd();
-            s += this.form.TableBodyStart();
-
-            return s;
-        }
-
-        public void AddSessionsFiles(bool scrub)
-        {
-            this.df.JobSessionInfoToXml(scrub);
         }
 
         public string AddConfigurationTablesHeader()
